+++ lib/scheduler/scheduler.h
@@ -14,7 +14,7 @@ extern "C" {
 */
 
 typedef struct {
-    int period;                     // period in seconds
+    double period;                     // period in seconds
     int delay;                      // seconds until next activation
     unsigned long nextExec;         // UNIX time of next activation
     void (*func)(void);             // function pointer
@@ -44,7 +44,7 @@ void Sched_Dispatch(void);
 * - void (*f)(void) - Read function;
 * - int p - period in seconds;
 */
-int Sched_AddT(void (*f)(void), int p);
+int Sched_AddT(void (*f)(void), double p);
 
 
 #ifdef __cplusplus
